pyproject.toml
[project]
name = "SYM-wallasls"
version = "0.0.0"
dependencies = [
  "aws-cdk-lib==2.179.0",
  "constructs>=10.0.0,<11.0.0",
  "confz==2.1.0"
]
requires-python = ">=3.11,<4"
authors = [
  {name = "Sergio Yunta Martín", email = "sergio.yuntam@gmail.com"}
]

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
aws-cdk-lib = "2.179.0"
constructs = "~10"
confz = "2.1.0"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
boto3 = "*"
shortuuid = "1.0.13"

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

line-length = 80
indent-width = 4

target-version = "py311"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = false
docstring-code-line-length = "dynamic"